Added planar control for arm
